rename "FreeTasks" to "AvailableMessageHandlers" I feel this is a better name as "FreeTasks" could be misread as an action "this method will free up all the tasks".
Also surfaced the "MaxConcurrentMessageHandlers" value as I want to log this at startup (to monitor what the DefaultThrottledThroughput is doing). You *can* use FreeTasks for that on a newly created IMessageProcessingStrategy on the assumption that no tasks are in use yet, so the current available count equals the max available count, but that's not great.

This is a minor breaking change (which probably affects mostly me)

And added comments to IMessageProcessingStrategy