An exception for when publishing fails (#260)
* An exception for when publishing fails
Include the relevant data
NB we retry publishing (3 times by default), so the exception seen outside of JustSaying will be from the last attempt. Previous exceptions will be discarded

* Log all message publishing failures
Even those that are going to be retried. Log as a warning with an exception. Similar wording as with giving up after last attempt.