Refactor test configuration
Refactor the tests' configuration to support changing the regions, keys etc. in a central place.
For example, a developer in Australia may wish to configure the tests locally to always point to ap-southeast-2.
Also refactors and renames some stub types to reduce duplication and confusion.
Move some classes to their own files.